Generally usable processes defined in keys.ts were moved to a separate module. In keys.ts, only processes called from router are defined.
