+++ common/src/keys.ts
@@ -1,5 +1,4 @@
-import { CRV, newPrivateJwk, PrivateJwk, PublicJwk } from "elliptic-jwk";
-import keyutil from "js-crypto-key-utils";
+import { CRV, newPrivateJwk, PublicJwk } from "elliptic-jwk";
 
 import { NotSuccessResult } from "./routerCommon";
 import { UNIQUE_CONSTRAINT_FAILED } from "./store.js";
@@ -12,13 +11,7 @@ import {
   createEcdsaCsr,
   createEcdsaSelfCertificate,
 } from "./x509.js";
-
-interface Csr {
-  csr: string;
-}
-interface X509Cert {
-  cert: string;
-}
+import { jwkToPem } from "./crypto/util";
 
 const INVALID_PARAMETER_ERROR: NgResult<NotSuccessResult> = {
   ok: false,
@@ -77,52 +70,6 @@ const isSupportedCurve = (value: string): Result<CRV, string> => {
         error: `Invalid value: ${value}. Allowed values are: secp256k1, Ed25519`,
       };
   }
-};
-
-interface PemKeyPair {
-  publicKey: string;
-  privateKey: string;
-}
-
-const curveNistName = (crv: string): string => {
-  switch (crv) {
-    case "secp256k1":
-      return "P-256K";
-    case "secp256r1":
-      return "P-256";
-    case "secp384r1":
-      return "P-384";
-    case "secp521r1":
-      return "P-521";
-    default:
-      throw new Error(`Unsupported curve: ${crv}`);
-  }
-};
-
-const jwkToPem = async (jwk: {
-  kty: "EC" | "OKP";
-  d: string;
-  crv: "P-256" | "secp256k1" | "Ed25519";
-  x: string;
-  y: string | undefined;
-}): Promise<PemKeyPair> => {
-  // See https://github.com/junkurihara/jscu/blob/8168ab947e23876d2915ed8849f021d59673e8aa/packages/js-crypto-key-utils/src/params.ts#L12
-  const preprocessedJwk = jwk.crv.startsWith("sec")
-    ? {
-        kty: jwk.kty,
-        d: jwk.d,
-        crv: curveNistName(jwk.crv),
-        x: jwk.x,
-        y: jwk.y,
-      }
-    : jwk;
-
-  const keyObj = new keyutil.Key("jwk", preprocessedJwk);
-  const privatePem = (await keyObj.export("pem")) as string;
-  const publicPem = (await keyObj.export("pem", {
-    outputPublic: true,
-  })) as string;
-  return { publicKey: publicPem, privateKey: privatePem };
 };
 
 export const genKey = async (
@@ -224,6 +171,13 @@ export const revokeKey = async (
     return UNKNOWN_ERROR;
   }
 };
+
+interface Csr {
+  csr: string;
+}
+interface X509Cert {
+  cert: string;
+}
 
 export const createCsr = async (
   keyId: string,
@@ -360,22 +314,6 @@ export const registerCert = async (
       return toInternalError(name, message);
     }
     return UNKNOWN_ERROR;
-  }
-};
-
-export const getKeyAlgorithm = (jwk: PrivateJwk): string => {
-  switch (jwk.kty) {
-    case "EC":
-      // todo add patterns of crv
-      if (jwk.crv === "P-256") {
-        return "ES256";
-      } else {
-        return "ES256K";
-      }
-    case "OKP":
-      return "EdDSA";
-    default:
-      throw new Error("Unsupported key type");
   }
 };
